feat: adding analyzer I/O abstractions

+++ src/analyzer/analyzer_new.c
@@ -11,7 +11,7 @@ int8_t analyzer_new(analyzer_t *self, const options_t *options) {
     return -1;
   }
   fseeko(self->binary, 0, SEEK_END);
-  self->limit = ftello(self->binary);
+  self->limit = (uint64_t)ftello(self->binary);
   fseeko(self->binary, 0, SEEK_SET);
   return 0;
 }
